feat: setting up gql queries with fragments
