+++ packages/headless/src/api/queries/index.ts
@@ -1,5 +1,5 @@
-export * from './GET_POSTS';
-export * from './GET_CONTENT_NODE';
+export * from './getPosts';
+export * from './getContentNode';
 export * from './GENERAL_SETTINGS';
 export * from './GET_URI_INFO';
 export * from './PAGE_DATA_FRAGMENT';
